Lots of cleanup to include the ...
correct source file, create a temporary canonical file to make the
script flow simpler
Retrieved the git copy of this file and make sure that the canonical
table is updated if this source file is newer than the canonical file.
Fixed the test for presence of the b3sum binary.  Fixed the case
statement to allow for upper or lower case replies. added a test for
verifying that the canonical source file is present.  It was previously
missing due to a bug, this tests should never succeed again.  Fixed
bumping the line count when reading the file.  There still may be a
problem with read_hashbits.  Fixed the computation of hexdigits. Fixed
typos in canonical_dir