multiple changes ...
Created a regular expression for a canonical hash number
Added a number of disagnostic statements that show the Associative
arrays are OK inside the while loop, but are suddenly empty after just
exiting the loop.  This is a mystery to me.
Removed a test for if a binary is found and just used the output of the
which command.
Removed " around associative array subscripts.