Modified the HERE file for the ...
canonical source, but it won't get picked up until
the modification date on this script is newer than
the last saved file.

Made the pushd/popd silent

Changed VERBOSE_USAGE to DEBUG_USAGE which was moved into func.regex
where it belongs.

Move the regular expression for canonical hash numbers into
func.hashcanonical where it belongs

pulled in yfunc.global to get the definition of YesFS dir, but added a
local definition if that source pull is removed.

testing of the debug flag is limited to a single decimal digit.

Added a comment to the end of the getopts loop

Added distinctive comments around deep diagnostics that should be
removed when we figure out why the global associative arrays are
misbehaving

dumped each line as we read from the canonical source file.

Added deep diagnostics to print out the keys and values for all of the
Associative Arrays.