I finally found the culprit.  It was ...
piping a command into a while loop.  The global associative arrays
modified in the while loop stayed with the sub shell for the piped
command and they were no longer available outside the while loop.  I had
to redirect the command to a file and then use that file as the input to
the while loop.

There is still some cleanup on creating the files, but that is
mechanical work to clean up.